OLDLIB: Initial commit of the old eos library
This is an old library for reading/writing NWN files that I worked on
between 2004 and 2005. The plan was to expand it into a full-fledged
portable toolset, but it never got that far.

Readers beware: The code is quite ugly. I only release it because
the tools provided by the test application, especially the GFF to
XML converter, still are useful. They should probably be replaced
by proper clean tools in the future.